refactor: move cli util to shared
